requirements.txt
django-tinymce
pillow